Fix build for pip with --use-pep517
Resolves #660

When building this package using `pip install --use-pep517 --no-binary :all:`
or in dev `pip install -e . --use-pep517` we see a failure in setup.py:

      [...]
      AttributeError: 'dict' object has no attribute '__LIGHTFM_SETUP__'

`__builtins__` is an implementation detail and:

    The value of __builtins__ is normally either this module or the
    value of this module’s __dict__ attribute

When using pip with `--use-pep517` it seems to be a dict, and otherwise
seems to be the `builtins` object.

Rather than switch to importing the `builtins` module I've taken the
opportunity to rework the package version code to follow approach (3)
detailed in https://packaging.python.org/en/latest/guides/single-sourcing-package-version/

This feels a little more forgiving than the current approach in that it
avoids loading the actual package itself in `setup.py`, skipping the
workaround for the fact LightFM relies on a native extension.